Rename namespace to match directories
